aider: Modify client-side JavaScript code to display server response in output div when register button is clicked.
# Aider chat conversation:

ASSISTANT: To display the server log in the output div, we need to modify the client-side JavaScript code in the `client/index.html` file to handle the response from the server when the register button is clicked.

Here is the change:

```html
client/index.html
<<<<<<< SEARCH
        function register() {
            // Register user in backend upon connecting
            fetch('/register', {
                method: 'POST',
                headers: {
                    'Content-Type': 'application/json',
                },
                body: JSON.stringify({}),
            }).then((res) => {
                console.log(res);
            }).catch((err) => {
                console.error(err);
            })
        }
=======
        function register() {
            // Register user in backend upon connecting
            fetch('/register', {
                method: 'POST',
                headers: {
                    'Content-Type': 'application/json',
                },
                body: JSON.stringify({}),
            })
            .then(response => response.json())
            .then(data => {
                console.log(data);
                document.getElementById('outputDiv').innerText = `Player ${data.playerId} has joined the game.\nSession ID: ${data.sessionId}\nPlayer ID: ${data.playerId}`;
            })
            .catch((err) => {
                console.error(err);
            })
        }
>>>>>>> REPLACE
```

This change will make the server response visible in the output div when the register button is clicked. The server response includes the player ID and the session ID.

Please note that this change assumes that the server sends the session ID and the player ID in the response body when the '/register' endpoint is hit. If the server does not send these data, the server code needs to be modified accordingly.